# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "研究室内の備品を管理するシステムです。",
        "title": "備品管理システム",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/disposal/all": {
            "get": {
                "description": "全ての資産廃棄履歴を取得します。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disposal"
                ],
                "summary": "全廃棄履歴の取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DisposalListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/disposal/{id}": {
            "get": {
                "description": "指定されたIDの廃棄履歴を取得します。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disposal"
                ],
                "summary": "廃棄履歴の取得 (ID指定)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "廃棄ID (Disposal ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DisposalResponseWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "指定された資産IDの資産を廃棄登録します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disposal"
                ],
                "summary": "資産の廃棄処理",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "資産ID (Asset ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "廃棄登録リクエスト",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateDisposalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/lend/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lend"
                ],
                "summary": "貸出情報一覧取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.LendListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/lend/all/with-name": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lend"
                ],
                "summary": "貸出情報一覧取得 (名称付き)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.LendingDetailListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/lend/edit/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lend"
                ],
                "summary": "貸出情報編集",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "貸出ID (Lend ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新情報",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.EditLendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/lend/return/{id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lend"
                ],
                "summary": "備品返却処理",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "貸出ID (Lend ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "返却情報",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ReturnAssetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/lend/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lend"
                ],
                "summary": "貸出情報取得 (ID指定)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "貸出ID (Lend ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.LendResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lend"
                ],
                "summary": "備品貸出処理",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "備品ID (Asset ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "貸出情報",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.LendAssetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.AssetsLend": {
            "type": "object",
            "properties": {
                "actual_return_date": {
                    "description": "実際の返却日 (\"YYYY-MM-DD\" or null)",
                    "type": "string"
                },
                "asset_id": {
                    "description": "貸出対象資産ID",
                    "type": "integer"
                },
                "borrower": {
                    "description": "借用者",
                    "type": "string"
                },
                "expected_return_date": {
                    "description": "返却予定日 (\"YYYY-MM-DD\" or null)",
                    "type": "string"
                },
                "id": {
                    "description": "主キー",
                    "type": "integer"
                },
                "lend_date": {
                    "description": "貸出日 (\"YYYY-MM-DD\")",
                    "type": "string"
                },
                "notes": {
                    "description": "備考 (null可)",
                    "type": "string"
                },
                "quantity": {
                    "description": "貸出数",
                    "type": "integer"
                }
            }
        },
...
        "domain.CreateDisposalRequest": {
            "type": "object",
            "properties": {
                "is_individual": {
                    "description": "個別管理かどうか",
                    "type": "boolean"
                },
                "processed_by": {
                    "description": "処理者の学籍番号",
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "domain.DisposalResponse": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "description": "対象資産 ID",
                    "type": "integer"
                },
                "disposal_date": {
                    "description": "廃棄日 (YYYY-MM-DD)",
                    "type": "string"
                },
                "id": {
                    "description": "主キー",
                    "type": "integer"
                },
                "processed_by": {
                    "description": "処理担当者",
                    "type": "string"
                },
                "quantity": {
                    "description": "廃棄数量",
                    "type": "integer"
                },
                "reason": {
                    "description": "廃棄理由",
                    "type": "string"
                }
            }
        },
...
        "domain.EditLendRequest": {
            "type": "object",
            "required": [
                "borrower",
                "lend_date",
                "quantity"
            ],
            "properties": {
                "actual_return_date": {
                    "description": "NULL 可（通常登録時はNULL）",
                    "type": "string"
                },
                "borrower": {
                    "type": "string"
                },
                "expected_return_date": {
                    "description": "NULL 可",
                    "type": "string"
                },
                "lend_date": {
                    "description": "\"YYYY-MM-DD\" 形式",
                    "type": "string"
                },
                "notes": {
                    "description": "NULL 可",
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "domain.LendAssetRequest": {
            "type": "object",
            "required": [
                "asset_id",
                "borrower",
                "lend_date",
                "quantity"
            ],
            "properties": {
                "actual_return_date": {
                    "description": "NULL 可（通常登録時はNULL）",
                    "type": "string"
                },
                "asset_id": {
                    "type": "integer"
                },
                "borrower": {
                    "type": "string"
                },
                "expected_return_date": {
                    "description": "NULL 可",
                    "type": "string"
                },
                "lend_date": {
                    "description": "\"YYYY-MM-DD\" 形式",
                    "type": "string"
                },
                "notes": {
                    "description": "NULL 可",
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "domain.LendingDetail": {
            "type": "object",
            "properties": {
                "borrower": {
                    "type": "string"
                },
                "expected_return_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lend_date": {
                    "type": "string"
                },
                "manufacturer": {
                    "type": "string"
                },
                "model_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "domain.ReturnAssetRequest": {
            "type": "object",
            "required": [
                "actual_return_date",
                "quantity"
            ],
            "properties": {
                "actual_return_date": {
                    "type": "string"
                },
                "lendID": {
                    "type": "integer"
                },
                "notes": {
                    "description": "NULL 可",
                    "type": "string"
                },
                "quantity": {
                    "description": "返却数量",
                    "type": "integer"
                }
            }
        },
...
        "handler.DisposalListResponse": {
            "type": "object",
            "properties": {
                "disposals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.DisposalResponse"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Disposals fetched successfully"
                }
            }
        },
        "handler.DisposalResponseWrapper": {
            "type": "object",
            "properties": {
                "disposal": {
                    "$ref": "#/definitions/domain.DisposalResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Disposal fetched successfully"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "不正なリクエストです"
                }
            }
        },
        "handler.LendListResponse": {
            "type": "object",
            "properties": {
                "lends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.AssetsLend"
                    }
                }
            }
        },
        "handler.LendResponse": {
            "type": "object",
            "properties": {
                "lend": {
                    "$ref": "#/definitions/domain.AssetsLend"
                },
                "message": {
                    "type": "string",
                    "example": "Fetch completed"
                }
            }
        },
        "handler.LendingDetailListResponse": {
            "type": "object",
            "properties": {
                "lends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.LendingDetail"
                    }
                }
            }
        },
...
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "処理に成功しました"
                }
            }
        }
    }
}